Solving several issues and errors
Returning reference to this instead of pointer (for eye-candy setup).
Fixing wrong setup of getters (using two methods instead of a single one).
Wrappers will take a TMultiGraph pointer as argument, no make sure the directory of the created graph is owned by the user and to be useable also as standalone methods.